refactor: create package for holding generation of driver samples.
