+++ src/tests/usecase/delete_driver_test.go
@@ -8,12 +8,13 @@ import (
 	repository "github.com/rafaft/truck-driver-trip-system/adapter/gateway/database"
 	"github.com/rafaft/truck-driver-trip-system/entity"
 	"github.com/rafaft/truck-driver-trip-system/infrastructure/log"
+	"github.com/rafaft/truck-driver-trip-system/tests/samples"
 	"github.com/rafaft/truck-driver-trip-system/usecase"
 )
 
 func TestDeleteDriver(t *testing.T) {
 	l := log.NewFakeLogger()
-	r := repository.NewDriverInMemory(getDriversSample(t))
+	r := repository.NewDriverInMemory(samples.GetDrivers(t))
 	uc := usecase.NewDeleteDriver(l, r)
 
 	tests := []struct {
