+++ src/tests/usecase/get_drivers_test.go
@@ -9,12 +9,13 @@ import (
 	repository "github.com/rafaft/truck-driver-trip-system/adapter/gateway/database"
 	"github.com/rafaft/truck-driver-trip-system/entity"
 	"github.com/rafaft/truck-driver-trip-system/infrastructure/log"
+	"github.com/rafaft/truck-driver-trip-system/tests/samples"
 	"github.com/rafaft/truck-driver-trip-system/usecase"
 )
 
 func TestGetDrivers(t *testing.T) {
 	l := log.NewFakeLogger()
-	r := repository.NewDriverInMemory(getDriversSample(t))
+	r := repository.NewDriverInMemory(samples.GetDrivers(t))
 	uc := usecase.NewGetDrivers(l, r)
 
 	tests := []struct {
