+++ src/tests/usecase/update_driver_test.go
@@ -9,6 +9,7 @@ import (
 	repository "github.com/rafaft/truck-driver-trip-system/adapter/gateway/database"
 	"github.com/rafaft/truck-driver-trip-system/entity"
 	"github.com/rafaft/truck-driver-trip-system/infrastructure/log"
+	"github.com/rafaft/truck-driver-trip-system/tests/samples"
 	"github.com/rafaft/truck-driver-trip-system/usecase"
 )
 
@@ -16,7 +17,7 @@ func TestUpdateDriver(t *testing.T) {
 	now := time.Now()
 
 	l := log.NewFakeLogger()
-	r := repository.NewDriverInMemory(getDriversSample(t))
+	r := repository.NewDriverInMemory(samples.GetDrivers(t))
 	uc := usecase.NewUpdateDriver(l, r)
 
 	tests := []struct {
